KAFKA-15164; [1/N] Extract reusable functionality from OffsetsForLeaderEpochClient for validating positions
